requirements.txt
feedparser
requests
beautifulsoup4
lxml
pyyaml
python-dateutil
ratelimit
playwright
